# "Double" the performance