Well, it uses **50% less power** - that's "double" performance. Basically, instead of using spinlocks, I made `whisper.cpp` use condition variables with mutices (mutex plural?).

Whisper is not really a low latency system. This means that busy locks aren't the best choice for synchronisation. The more so, that `whisper.cpp` is also supposed to run on the web and on mobile devices, where users usually care about power usage. In this PR, I made `whisper.cpp` use the classical conditional variable + mutex lock schema instead. On a 12900KS without overclocking, this reduces the CPU usage (and hence the power consumption) by half. On the other hand, if we go for full 100% utilization, the computation time is reduced by about **25%**. Performance tables below.

This is a draft because I haven't implemented the lock using `pthreads` yet, and the current Windows implementation is rather naive and suboptimal. I am also yet to optimize the computations themselves.