@janekb04 2023-03-26
## Original version, 24 threads (95% utilization)
```
Running ggml_mul_mat benchmark with 24 threads

ggml_mul_mat:    64 x    64: F16      0.3 GFLOPS (128 runs) / F32      0.2 GFLOPS (128 runs)
ggml_mul_mat:   128 x   128: F16      2.7 GFLOPS (128 runs) / F32      1.5 GFLOPS (128 runs)
ggml_mul_mat:   256 x   256: F16     10.1 GFLOPS (128 runs) / F32     19.1 GFLOPS (128 runs)
ggml_mul_mat:   512 x   512: F16     61.7 GFLOPS (128 runs) / F32     74.5 GFLOPS (128 runs)
ggml_mul_mat:  1024 x  1024: F16    138.7 GFLOPS ( 65 runs) / F32    162.6 GFLOPS ( 76 runs)
ggml_mul_mat:  2048 x  2048: F16    184.7 GFLOPS ( 11 runs) / F32    192.5 GFLOPS ( 12 runs)
ggml_mul_mat:  4096 x  4096: F16    174.4 GFLOPS (  3 runs) / F32     94.8 GFLOPS (  3 runs)
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | tiny | 24 | 98 | 374 | fd83fb2 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | base | 24 | 153 | 1023 | fd83fb2 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | small | 24 | 437 | 2896 | fd83fb2 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | medium | 24 | 1301 | 8510 | fd83fb2 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | large | 24 | 2563 | 16643 | fd83fb2 |

## New version, 24 threads (50% utilization)
```
Running ggml_mul_mat benchmark with 24 threads

ggml_mul_mat:    64 x    64: F16      0.4 GFLOPS (128 runs) / F32      0.4 GFLOPS (128 runs)
ggml_mul_mat:   128 x   128: F16      3.0 GFLOPS (128 runs) / F32      2.9 GFLOPS (128 runs)
ggml_mul_mat:   256 x   256: F16     21.4 GFLOPS (128 runs) / F32     20.3 GFLOPS (128 runs)
ggml_mul_mat:   512 x   512: F16     95.0 GFLOPS (128 runs) / F32    100.3 GFLOPS (128 runs)
ggml_mul_mat:  1024 x  1024: F16    180.4 GFLOPS ( 84 runs) / F32    203.9 GFLOPS ( 95 runs)
ggml_mul_mat:  2048 x  2048: F16    207.3 GFLOPS ( 13 runs) / F32    179.8 GFLOPS ( 11 runs)
ggml_mul_mat:  4096 x  4096: F16    182.9 GFLOPS (  3 runs) / F32    107.1 GFLOPS (  3 runs)
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | tiny | 24 | 97 | 324 | de49899 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | base | 24 | 158 | 689 | de49899 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | small | 24 | 437 | 2384 | de49899 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | medium | 24 | 1301 | 8923 | de49899 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | large | 24 | 2540 | 16748 | de49899 |

## Old version, 120 threads (99% utilization)
```
Running ggml_mul_mat benchmark with 120 threads

ggml_mul_mat:    64 x    64: F16      0.0 GFLOPS (  3 runs) / F32      0.0 GFLOPS (  3 runs)
ggml_mul_mat:   128 x   128: F16      0.0 GFLOPS (  3 runs) / F32      0.0 GFLOPS (  3 runs)
ggml_mul_mat:   256 x   256: F16      0.0 GFLOPS (  3 runs) / F32      0.0 GFLOPS (  3 runs)
ggml_mul_mat:   512 x   512: F16      0.1 GFLOPS (  3 runs) / F32      0.2 GFLOPS (  3 runs)
ggml_mul_mat:  1024 x  1024: F16      1.6 GFLOPS (  3 runs) / F32      1.2 GFLOPS (  3 runs)
ggml_mul_mat:  2048 x  2048: F16     12.3 GFLOPS (  3 runs) / F32     12.6 GFLOPS (  3 runs)
ggml_mul_mat:  4096 x  4096: F16     68.5 GFLOPS (  3 runs) / F32     50.5 GFLOPS (  3 runs)
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | tiny | 120 | 96 | 78836 | fd83fb2 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | base | 120 | 157 | 113952 | fd83fb2 |
A while it took, indeed.

## New version, 120 threads (90% utilization)
```
Running ggml_mul_mat benchmark with 120 threads

ggml_mul_mat:    64 x    64: F16      0.1 GFLOPS (106 runs) / F32      0.1 GFLOPS (106 runs)
ggml_mul_mat:   128 x   128: F16      0.4 GFLOPS (106 runs) / F32      0.4 GFLOPS (104 runs)
ggml_mul_mat:   256 x   256: F16      3.5 GFLOPS (106 runs) / F32      3.5 GFLOPS (104 runs)
ggml_mul_mat:   512 x   512: F16     25.1 GFLOPS ( 94 runs) / F32     25.7 GFLOPS ( 96 runs)
ggml_mul_mat:  1024 x  1024: F16    129.0 GFLOPS ( 61 runs) / F32    127.7 GFLOPS ( 60 runs)
ggml_mul_mat:  2048 x  2048: F16    248.5 GFLOPS ( 15 runs) / F32    179.5 GFLOPS ( 11 runs)
ggml_mul_mat:  4096 x  4096: F16    191.5 GFLOPS (  3 runs) / F32    121.3 GFLOPS (  3 runs)
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | tiny | 120 | 98 | 583 | 071f410 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | base | 120 | 158 | 972 | 071f410 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | small | 120 | 435 | 2588 | 071f410 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | medium | 120 | 1296 | 7457 | 071f410 |
| 12th Gen Intel(R) Core(TM) i9-12900KS | Microsoft Windows 11 Pro | AVX2 | large | 120 | 2540 | 12715 | 071f410 |