@janekb04 2023-03-26
I also added a script that automatically runs all benchmarks on Windows. It is simply the shell script version converted into Powershell.