@janekb04 2023-03-26
And here's the `pthread` version. Now this should be mergeable, though as I wrote, I am planning further optimizations. macOS tables below.