@janekb04 2023-03-26
## Original version, 6 threads (75% utilization)
```
Running ggml_mul_mat benchmark with 6 threads

ggml_mul_mat:    64 x    64: F16      6.0 GFLOPS (128 runs) / F32      5.2 GFLOPS (128 runs)
ggml_mul_mat:   128 x   128: F16     66.8 GFLOPS (128 runs) / F32     42.1 GFLOPS (128 runs)
ggml_mul_mat:   256 x   256: F16    356.6 GFLOPS (128 runs) / F32    283.2 GFLOPS (128 runs)
ggml_mul_mat:   512 x   512: F16    459.2 GFLOPS (128 runs) / F32    530.5 GFLOPS (128 runs)
ggml_mul_mat:  1024 x  1024: F16    937.4 GFLOPS (128 runs) / F32   1379.4 GFLOPS (128 runs)
ggml_mul_mat:  2048 x  2048: F16   1217.2 GFLOPS ( 71 runs) / F32   1557.5 GFLOPS ( 91 runs)
ggml_mul_mat:  4096 x  4096: F16   1695.6 GFLOPS ( 13 runs) / F32   1431.4 GFLOPS ( 11 runs)
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | tiny | 6 | 49 | 106 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | base | 6 | 64 | 196 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | small | 6 | 178 | 674 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | medium | 6 | 558 | 1940 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | large | 6 | 1246 | 3547 | fd83fb2 |

## Original version, 10 threads (90% utilization)
```
Running ggml_mul_mat benchmark with 10 threads

ggml_mul_mat:    64 x    64: F16      3.7 GFLOPS (128 runs) / F32      3.0 GFLOPS (128 runs)
ggml_mul_mat:   128 x   128: F16     35.1 GFLOPS (128 runs) / F32     19.5 GFLOPS (128 runs)
ggml_mul_mat:   256 x   256: F16    223.6 GFLOPS (128 runs) / F32    130.8 GFLOPS (128 runs)
ggml_mul_mat:   512 x   512: F16    597.1 GFLOPS (128 runs) / F32    574.2 GFLOPS (128 runs)
ggml_mul_mat:  1024 x  1024: F16    773.1 GFLOPS (128 runs) / F32    528.9 GFLOPS (128 runs)
ggml_mul_mat:  2048 x  2048: F16    599.6 GFLOPS ( 35 runs) / F32    485.6 GFLOPS ( 29 runs)
ggml_mul_mat:  4096 x  4096: F16   1005.8 GFLOPS (  8 runs) / F32    722.6 GFLOPS (  6 runs)
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | tiny | 10 | 46 | 131 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | base | 10 | 65 | 286 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | small | 10 | 180 | 1105 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | medium | 10 | 526 | 3225 | fd83fb2 |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | large | 10 | 1237 | 5546 | fd83fb2 |

## New version, 6 threads (45% utilization)
```
Running ggml_mul_mat benchmark with 10 threads

[deadlock of some kind, I'll have to look into this]
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | tiny | 6 | 49 | 121 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | base | 6 | 66 | 203 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | small | 6 | 175 | 667 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | medium | 6 | 507 | 1818 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | large | 6 | 1315 | 3249 | a6ee46f |

## New version, 10 threads (50% utilization)
```
Running ggml_mul_mat benchmark with 10 threads

[deadlock of some kind, I'll have to look into this]
```
| CPU | OS | Config | Model | Th | Load | Enc. | Commit |
| --- | -- | ------ | ----- | -- | ---- | ---- | ------ |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | tiny | 10 | 42 | 129 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | base | 10 | 67 | 240 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | small | 10 | 176 | 744 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | medium | 10 | 545 | 1933 | a6ee46f |
| Apple M1 Pro | macOS 13.2.1 (22D68) |  NEON BLAS | large | 10 | 1419 | 3347 | a6ee46f |